Renamed grpc client to example
